requirements.txt
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
opencv-python==4.2.0.32
pandas==1.3.0
Pillow==8.3.2
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.4.0
torchvision==0.5.0
typing==3.7.4.3
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
